# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngredientDisplayItem": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "name",
          "display",
          "amount",
          "unit"
        ],
        "title": "IngredientDisplayItem",
        "type": "object"
      },
...
      "LikeStatus": {
        "properties": {
          "like_count": {
            "title": "Like Count",
            "type": "integer"
          },
          "liked": {
            "title": "Liked",
            "type": "boolean"
          }
        },
        "required": [
          "liked",
          "like_count"
        ],
        "title": "LikeStatus",
        "type": "object"
      },
...
      "MySubmissionDetail": {
        "properties": {
          "calorie": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calorie"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "ingredient_name": {
            "title": "Ingredient Name",
            "type": "string"
          },
          "price_per_100g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per 100G"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "sodium_mg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sodium Mg"
          }
        },
        "required": [
          "ingredient_name",
          "calorie",
          "carbs_g",
          "protein_g",
          "fat_g",
          "sodium_mg",
          "price_per_100g"
        ],
        "title": "MySubmissionDetail",
        "type": "object"
      },
      "MySubmissionItem": {
        "properties": {
          "calorie": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calorie"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ingredient_name": {
            "title": "Ingredient Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ingredient_name",
          "calorie",
          "status"
        ],
        "title": "MySubmissionItem",
        "type": "object"
      },
...
      "PopularRecipeItem": {
        "properties": {
          "calorie": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calorie"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "like_count": {
            "title": "Like Count",
            "type": "integer"
          },
          "menu_name": {
            "title": "Menu Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "menu_name",
          "category",
          "calorie",
          "image_url",
          "like_count"
        ],
        "title": "PopularRecipeItem",
        "type": "object"
      },
...
      "SubmissionRequest": {
        "properties": {
          "calorie": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calorie"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "ingredient_name": {
            "title": "Ingredient Name",
            "type": "string"
          },
          "price_per_100g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per 100G"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "sodium_mg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sodium Mg"
          }
        },
        "required": [
          "ingredient_name"
        ],
        "title": "SubmissionRequest",
        "type": "object"
      },
      "SubmissionResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          }
        },
        "required": [
          "status"
        ],
        "title": "SubmissionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "냉장고 한끼 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ingredient-submissions": {
      "get": {
        "operationId": "list_my_submissions_ingredient_submissions_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MySubmissionItem"
                  },
                  "title": "Response List My Submissions Ingredient Submissions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Submissions",
        "tags": [
          "ingredient-submission"
        ]
      },
      "post": {
        "operationId": "submit_ingredient_ingredient_submissions_post",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Ingredient",
        "tags": [
          "ingredient-submission"
        ]
      }
    },
    "/ingredient-submissions/{submission_id}": {
      "get": {
        "operationId": "get_submission_detail_ingredient_submissions__submission_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MySubmissionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Submission Detail",
        "tags": [
          "ingredient-submission"
        ]
      },
      "put": {
        "operationId": "update_submission_ingredient_submissions__submission_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Submission",
        "tags": [
          "ingredient-submission"
        ]
      }
    },
...
    "/recommendation/recipes/popular": {
      "get": {
        "description": "홈의 \"많이 추천한 메뉴\" 줄 - 로그인 여부와 무관하게 볼 수 있는 공개 정보라\n인가를 요구하지 않는다(search_all_recipes와 동일한 방침).",
        "operationId": "get_popular_recipes_recommendation_recipes_popular_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PopularRecipeItem"
                  },
                  "title": "Response Get Popular Recipes Recommendation Recipes Popular Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Popular Recipes",
        "tags": [
          "like"
        ]
      }
    },
...
    "/recommendation/recipes/{recipe_id}/ingredients": {
      "get": {
        "description": "가구원 수에 맞춰 환산된 재료 수량 목록. portion_agent.py 로직 그대로 재사용한다.",
        "operationId": "get_recipe_ingredients_scaled_recommendation_recipes__recipe_id__ingredients_get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "title": "Recipe Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientDisplayItem"
                  },
                  "title": "Response Get Recipe Ingredients Scaled Recommendation Recipes  Recipe Id  Ingredients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recipe Ingredients Scaled",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/recommendation/recipes/{recipe_id}/like": {
      "get": {
        "operationId": "get_like_status_recommendation_recipes__recipe_id__like_get",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "title": "Recipe Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Like Status",
        "tags": [
          "like"
        ]
      }
    },
    "/recommendation/recipes/{recipe_id}/like/toggle": {
      "post": {
        "operationId": "toggle_like_recommendation_recipes__recipe_id__like_toggle_post",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "title": "Recipe Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Like",
        "tags": [
          "like"
        ]
      }
    },
...
  }
}